requirements.txt
websockets==12.0
aiohttp==3.9.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pandas==2.1.4
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0
requests==2.31.0
